style(core-components-typography): rename to lower case
