+++ packages/amount/src/component.stories.mdx
@@ -4,7 +4,7 @@ import { Container, Row, Col } from 'storybook/blocks/grid';
 import { ComponentHeader } from 'storybook/blocks/component-header';
 
 import { Amount } from '.';
-import { Typography } from '../../typography/src/Component';
+import { Typography } from '../../typography/src';
 import { getAllCurrencyCodes } from '@alfalab/utils';
 import { name, version } from '../package.json';
