+++ packages/typography/src/component.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, Props, Preview } from '@storybook/addon-docs/blocks';
 import { text, boolean, select, number } from '@storybook/addon-knobs';
 import { ComponentHeader } from 'storybook/blocks/component-header';
 
-import { Typography } from './Component';
+import { Typography } from './component';
 import { name, version } from '../package.json';
 import { colors } from './colors';
