+++ packages/typography/src/component.tsx
