+++ packages/typography/src/text/component.tsx
