+++ packages/typography/src/text/index.ts
@@ -1,1 +1,1 @@
-export * from './Component';
+export * from './component';
